Change test execution order and add async tests to async test file
