+++ pyscriptjs/tests/integration/test_02_output.py
@@ -245,40 +245,3 @@ class TestOutuput(PyScriptTest):
         console_text = self.console.all.lines
         assert console_text.index("1print") == (console_text.index("2print") - 1)
         assert console_text.index("1console") == (console_text.index("2console") - 1)
-
-    def test_multiple_async_display(self):
-        self.pyscript_run(
-            """
-                <py-script id="py1">
-                def say_hello():
-                    display('hello')
-                </py-script>
-                <py-script id="py2">
-                    say_hello()
-                </py-script>
-            """
-        )
-        inner_html = self.page.content()
-        pattern = r'<div id="py2-2">hello</div>'
-        assert re.search(pattern, inner_html)
-
-    def test_multiple_async_multiple_display(self):
-        self.pyscript_run(
-            """
-                <py-script id='pyA'>
-                    import asyncio
-                    for i in range(2):
-                        display('A')
-                        await asyncio.sleep(0)
-                </py-script>
-
-                <py-script id='pyB'>
-                    import asyncio
-                    for i in range(2):
-                        display('B')
-                        await asyncio.sleep(0)
-                </py-script>
-            """
-        )
-        inner_text = self.page.inner_text("html")
-        assert "A\nB\nA\nB" in inner_text
